requirements.txt
chainlit==1.3.0rc1
openai
yfinance
websockets
pydantic==2.10.1
python-dotenv
atlassian-python-api
pyodbc
azure-search-documents